+++ Sources/Models/Point.swift
@@ -94,7 +94,7 @@ extension Point: Mappable {
     
     convenience init?(fromElement element: AEXMLElement) {
         // When the element is an error, don't create the instance.
-        if element.errored {
+        if let _ = element.error {
             return nil
         }