+++ Sources/Models/Route.swift
@@ -42,7 +42,7 @@ extension Route: Mappable {
     
     convenience init?(fromElement element: AEXMLElement) {
         // When the element is an error, don't create the instance.
-        if element.errored {
+        if let _ = element.error {
             return nil
         }