+++ Sources/Models/Track.swift
@@ -45,7 +45,7 @@ extension Track: Mappable {
     
     convenience init?(fromElement element: AEXMLElement) {
         // When the element is an error, don't create the instance.
-        if element.errored {
+        if let _ = element.error {
             return nil
         }