+++ Sources/Models/TrackSegment.swift
@@ -21,7 +21,7 @@ extension TrackSegment: Mappable {
     
     convenience init?(fromElement element: AEXMLElement) {
         // When the element is an error, don't create the instance.
-        if element.errored {
+        if let _ = element.error {
             return nil
         }