+++ Sources/Parser/Extensions/AEXMLElement+Parsing.swift
@@ -12,28 +12,24 @@ import AEXML
 extension AEXMLElement {
     
     var optionalStringValue: String? {
-        if errored {
+        if let _ = error {
             return nil
         }
         return value
     }
     
     var optionalIntValue: Int? {
-        if errored {
+        if let _ = error {
             return nil
         }
         return intValue
     }
     
     var optionalFloatValue: Float? {
-        if errored {
+        if let _ = error {
             return nil
         }
         return Float(doubleValue)
-    }
-    
-    var errored: Bool {
-        return name == AEXMLElement.errorElementName
     }
     
 }