+++ Package.swift
@@ -1,7 +1,7 @@
 import PackageDescription
 
 let package = Package(
-  name: "GPXKit",
+  name: "LogKit",
   dependencies: [
     .Package(url: "https://github.com/fousa/AEXML.git", majorVersion: 2, minor: 1)
   ]