+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# GPXKit
+# LogKit
 
 ## Version 1.1.0