+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # LogKit
 
+## Version 1.2.0
+
+- Rename `GPXKit` to `LogKit` because we want to be able to support multiple formats.
+- Update the code to be Swift 3.0 compatible.
+
 ## Version 1.1.0
 
 - Remove the empty parse error notice.